int unsigned int fix
git-svn-id: svn://84.194.112.164/bengine@356 2ee6282c-0497-4b59-8b7c-bcac2105c061

+++ src/gui/widget.cpp
@@ -32,7 +32,7 @@ void Widget::registerWidget( const string& name, Widget* nwidget )
 
 bool Widget::mouseOver(int x, int y)
 {
-	if ( active && x > position.x && x < position.x+*v_widthP && y > position.y && y < position.y+*v_heightP )
+	if ( active && x > position.x && x < position.x+(int)*v_widthP && y > position.y && y < position.y+(int)*v_heightP )
 		return true;
 	return false;
 }
